Refactor trait resolution: introduce TraitEnv, TraitContext, and extract trait_query
Restructure the resolver's trait resolution system for better maintainability:

Phase 1 - TraitEnv: Consolidate three Arc'd index fields (trait_impl_index,
trait_decl_index, blanket_trait_impl_index) into a single Arc<TraitEnv> struct
that represents immutable global trait knowledge built once per compilation.

Phase 2 - TraitContext: Replace four separate mutable fields (current_type_params,
current_type_param_bounds, current_associated_type_bindings, current_self_type)
with a single TraitContext value type. Multi-field save/restore patterns are
consolidated from multiple old_* variables into single saved_trait_ctx snapshots.

Phase 3 - trait_query.rs: Extract trait query functions (type_implements_trait,
find_trait_impl_for_type, check_impl_block_bounds, find_method_in_trait_bounds,
register_assoc_types_for_concrete_type_and_trait, etc.) from method_lookup.rs
into a dedicated module. Reduces method_lookup.rs from 3416 to 2591 lines.
