+++ wado-compiler/src/resolver/orchestration.rs
@@ -424,8 +424,7 @@ impl<'a, H: CompilerHost> Resolver<'a, H> {
         // Build trait lookup indices once for all modules.
         // This allows find_trait_method_for_type and find_indexing_trait_impl to do O(1)
         // lookups by type name instead of scanning all items in all modules per method call.
-        let (trait_impl_index, trait_decl_index, blanket_trait_impl_index) =
-            Self::build_trait_indices(modules);
+        let trait_env = super::trait_env::TraitEnv::build(modules);
 
         // Second pass: resolve each module with per-module function_return_types and imports
         for module_source in &sorted_sources {
@@ -545,24 +544,19 @@ impl<'a, H: CompilerHost> Resolver<'a, H> {
                 logger,
                 current_module_source: ModuleSource::entry_point_with_filename("<uninitialized>"), // Set in resolve_module
                 current_module_items: Vec::new(), // Set in resolve_module
-                current_type_params: IndexMap::default(),
-                current_type_param_bounds: IndexMap::default(),
+                trait_ctx: super::trait_env::TraitContext::default(),
                 generic_struct_names: IndexSet::default(),
                 generic_function_params: IndexMap::default(),
                 generic_function_resolved_param_types: IndexMap::default(),
                 generic_method_params: IndexMap::default(),
                 generic_method_resolved_param_types: IndexMap::default(),
-                current_associated_type_bindings: IndexMap::default(),
-                current_self_type: None,
                 wasi_registry,
                 builtin_registry: &builtin_registry,
                 current_module_globals: IndexMap::default(),
                 imported_globals: IndexMap::default(),
                 associated_constants: IndexMap::default(),
                 module_type_maps_cache: IndexMap::default(),
-                trait_impl_index: Arc::clone(&trait_impl_index),
-                trait_decl_index: Arc::clone(&trait_decl_index),
-                blanket_trait_impl_index: Arc::clone(&blanket_trait_impl_index),
+                trait_env: Arc::clone(&trait_env),
                 included_files,
                 known_type_names_cache: IndexSet::default(),
                 indexing_trait_cache: IndexMap::default(),
